+++ lib/locales/czech_converter.dart
@@ -1,7 +1,10 @@
 import '../enum/local_case_language_enum.dart';
 import 'number_converter.dart';
 
+/// A converter class for handling Czech-specific case conversions
+/// and number-to-words transformation.
 class CzechConverter {
+  /// A map for converting uppercase characters to lowercase specific to the Czech language.
   final Map<String, String> czechLowercaseMap = {
     'Á': 'á',
     'Č': 'č',
@@ -20,6 +23,7 @@ class CzechConverter {
     'Ž': 'ž'
   };
 
+  /// A map for converting lowercase characters to uppercase specific to the Czech language.
   final Map<String, String> czechUppercaseMap = {
     'á': 'Á',
     'č': 'Č',
@@ -38,6 +42,16 @@ class CzechConverter {
     'ž': 'Ž'
   };
 
+  /// Converts the given [text] to lowercase using Czech-specific rules.
+  ///
+  /// This function maps uppercase Czech characters to their lowercase
+  /// equivalents and returns the resulting lowercase string.
+  ///
+  /// Example:
+  /// ```dart
+  /// CzechConverter().toLowercase("ČÁSTKA");
+  /// // Output: "částka"
+  /// ```
   String toLowercase(String text) {
     return text
         .split('')
@@ -48,6 +62,16 @@ class CzechConverter {
         .toLowerCase();
   }
 
+  /// Converts the given [text] to uppercase using Czech-specific rules.
+  ///
+  /// This function maps lowercase Czech characters to their uppercase
+  /// equivalents and returns the resulting uppercase string.
+  ///
+  /// Example:
+  /// ```dart
+  /// CzechConverter().toUppercase("částka");
+  /// // Output: "ČÁSTKA"
+  /// ```
   String toUppercase(String text) {
     return text
         .split('')
@@ -58,6 +82,17 @@ class CzechConverter {
         .toUpperCase();
   }
 
+  /// Converts a given [number] to its word representation in Czech.
+  ///
+  /// This method uses Czech language-specific words for numbers and
+  /// supports values up to one million. For example, 256 is converted
+  /// to "dvě stě padesát šest" in Czech.
+  ///
+  /// Example:
+  /// ```dart
+  /// CzechConverter.convertNumberToWords(256);
+  /// // Output: "dvě stě padesát šest"
+  /// ```
   static String convertNumberToWords(int number) {
     final czechUnits = [
       "",