+++ lib/locales/german_converter.dart
@@ -1,7 +1,10 @@
 import '../enum/local_case_language_enum.dart';
 import 'number_converter.dart';
 
+/// A converter class for handling German-specific case conversions
+/// and number-to-words transformation.
 class GermanConverter {
+  /// A map for converting uppercase German characters to their lowercase equivalents.
   final Map<String, String> germanLowercaseMap = {
     'Ä': 'ä',
     'Ö': 'ö',
@@ -9,6 +12,7 @@ class GermanConverter {
     'ẞ': 'ß'
   };
 
+  /// A map for converting lowercase German characters to their uppercase equivalents.
   final Map<String, String> germanUppercaseMap = {
     'ä': 'Ä',
     'ö': 'Ö',
@@ -16,6 +20,16 @@ class GermanConverter {
     'ß': 'SS'
   };
 
+  /// Converts a number between 11 and 19 to its German word equivalent.
+  ///
+  /// This function provides specific words for German "teens" (11 to 19),
+  /// such as "elf" for 11 and "zwölf" for 12.
+  ///
+  /// Example:
+  /// ```dart
+  /// GermanConverter.germanTeens(13);
+  /// // Output: "dreizehn"
+  /// ```
   static String germanTeens(int number) {
     switch (number) {
       case 11:
@@ -41,6 +55,16 @@ class GermanConverter {
     }
   }
 
+  /// Converts the given [text] to lowercase using German-specific rules.
+  ///
+  /// This function maps uppercase German characters to their lowercase
+  /// equivalents and returns the resulting lowercase string.
+  ///
+  /// Example:
+  /// ```dart
+  /// GermanConverter().toLowercase("ÄPFEL");
+  /// // Output: "äpfel"
+  /// ```
   String toLowercase(String text) {
     return text
         .split('')
@@ -51,6 +75,16 @@ class GermanConverter {
         .toLowerCase();
   }
 
+  /// Converts the given [text] to uppercase using German-specific rules.
+  ///
+  /// This function maps lowercase German characters to their uppercase
+  /// equivalents and returns the resulting uppercase string.
+  ///
+  /// Example:
+  /// ```dart
+  /// GermanConverter().toUppercase("äpfel");
+  /// // Output: "ÄPFEL"
+  /// ```
   String toUppercase(String text) {
     return text
         .split('')
@@ -61,6 +95,17 @@ class GermanConverter {
         .toUpperCase();
   }
 
+  /// Converts a given [number] to its word representation in German.
+  ///
+  /// This method uses German language-specific words for numbers and
+  /// supports values up to one million. For example, 256 is converted
+  /// to "zweihundertsechsundfünfzig" in German.
+  ///
+  /// Example:
+  /// ```dart
+  /// GermanConverter.convertNumberToWords(256);
+  /// // Output: "zweihundertsechsundfünfzig"
+  /// ```
   static String convertNumberToWords(int number) {
     final germanUnits = [
       "",